CORE-356 code polishing
- HealthDepartments can be used directly as a service.
